requirements.txt
Flask==2.3.3
requests==2.31.0
python-gitlab==3.15.0
Werkzeug==2.3.7
redis==5.0.1
python-dotenv==1.0.0
pymongo==4.14.1
